==> 01_datasets_cleaning.ipynb <==
## Imports

import pandas as pd
import numpy as np

import re

df1 = pd.read_csv('all_poems.csv')
df2 = pd.read_csv('apcd.csv')

df2.head()

df1.head()

df2.shape

## Preprocessing

arabic_diacritics = re.compile(""" ّ    | # Tashdid
                             َ    | # Fatha
                             ً    | # Tanwin Fath
                             ُ    | # Damma
                             ٌ    | # Tanwin Damm
                             ِ    | # Kasra
                             ٍ    | # Tanwin Kasr
                             ْ    | # Sukun
                             ـ     # Tatwil/Kashida
                         """, re.VERBOSE)

def remove_diacritics(text):
    text = re.sub(arabic_diacritics, '', str(text))
    return text

clean_poems = df1['poem_text']
clean_apcd = df2['البيت'].apply(remove_diacritics)

## Export

clean_poems.to_csv('all_poems_clean.csv', index=False)
clean_apcd.to_csv('apcd_clean.csv', index=False)

<h3 style="color: red;">Note</h3>

    Datasets were not merged because APCD dataset is too large, and should probably include the first dataset already.
