==> 02_tokenization_embedding.ipynb <==
## Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from itertools import islice

import pickle as pkl

import sentencepiece as sp

import tkseem as tk
import gensim

df = pd.read_csv('data/apcd_clean.csv')
df['poem_text'] = df['البيت'].astype(str)
df.drop('البيت', inplace=True, axis=1)
df.head()

## Morphological Tokenizer

morph_tokenizer = tk.MorphologicalTokenizer(vocab_size=500_000)

morph_tokenizer.train()

text = 'توكلت في رزقي على الله خالقي'

morph_tokenizer.tokenize(text)

def tokenize_text_morph(text):
    return morph_tokenizer.tokenize(text)

df['morph_t'] = df.poem_text.apply(tokenize_text_morph)
df.head()

text = 'أق'

morph_tokenizer.tokenize(text)

morph_model = gensim.models.Word2Vec(
    df['morph_t'],
    window=5,
    min_count=4,
    workers=8)

morph_model.wv.most_similar(positive='ثلج')

morph_model.save('models/morph_word2vec_model.model')

## Word Tokenizer

word_tokenizer = tk.WordTokenizer()

df['poem_text'].to_csv('data.txt', sep=' ', index=False, header=False)
word_tokenizer.train('data.txt')

def tokenize_text_word(text):
    return word_tokenizer.tokenize(text)

df['word_t'] = df.poem_text.apply(tokenize_text_word)
df.head()

word_model = gensim.models.Word2Vec(
    df['word_t'],
    window=5,
    min_count=4,
    workers=8)

word_model.wv.most_similar(positive='جمال')

word_model.save('models/word_word2vec_model.model')

## Sentence Piece Tokenization

sp.SentencePieceTrainer.train(
    sentence_iterator=iter(df['poem_text'].values),
    model_prefix='models/sp_poem_model',
    vocab_size=50_000,
    pad_id=0,
    unk_id=1,
    bos_id=2,
    eos_id=3
)

spp = sp.SentencePieceProcessor(model_file='models/sp_poem_model.model')

def tokenize_text_sp(text):
    return spp.encode(text, out_type=str)

df['sp_t'] = df.poem_text.apply(tokenize_text_sp)
df.head()

sp_model = gensim.models.Word2Vec(
    df['sp_t'],
    window=5,
    min_count=3,
    workers=8)

sp_model.wv.most_similar(positive='أسد')

sp_model.save('models/sp_word2vec_model.model')

## Save Updated Data

df.to_csv('all_poems_tokenized.csv', index=False)
