==> 03_semantic_search.ipynb <==
## Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from itertools import islice

import sentencepiece as sp

from sklearn.metrics.pairwise import cosine_similarity
import arabicstopwords.arabicstopwords as stp

import tkseem as tk
import gensim

df = pd.read_csv('all_poems_tokenized.csv')
df.head()

morph_embedder = gensim.models.Word2Vec.load("models/morph_word2vec_model.model")
word_embedder = gensim.models.Word2Vec.load("models/word_word2vec_model.model")
sp_embedder = gensim.models.Word2Vec.load("models/sp_word2vec_model.model")
bert_embedder = gensim.models.Word2Vec.load("models/bert_word2vec_model.model")
bert_doc_embedder = gensim.models.Doc2Vec.load("models/bert_doc2vec_model.model")

## Sentence Embedding - Old

### morph_t

def morph_embed(tokens_poem):
    clear_poem = [token for token in tokens_poem if token != '<UNK>']
    embedded_tokens = np.array([morph_embedder.wv.__getitem__(token) for token in clear_poem if token in morph_embedder.wv.key_to_index])
    return np.mean(embedded_tokens, axis=0)

df['morph_embd'] = df['morph_t'].apply(morph_embed)
df['morph_embd']

### word_t

def word_embed(tokens_poem):
    clear_poem = [token for token in tokens_poem if token != '<UNK>']
    embedded_tokens = np.array([word_embedder.wv.__getitem__(token) for token in clear_poem if token in word_embedder.wv.key_to_index])
    return np.mean(embedded_tokens, axis=0)

df['word_embd'] = df['word_t'].apply(word_embed)
df['word_embd']

df['word_embd'].isna().sum()

### sp_t

def word_embed(tokens_poem):
    clear_poem = [token for token in tokens_poem if token not in ['-0-', '-1-', '-2-', '-3-']]
    embedded_tokens = np.array([sp_embedder.wv.__getitem__(token) for token in clear_poem if token in sp_embedder.wv.key_to_index])
    return np.mean(embedded_tokens, axis=0)

df['sp_embd'] = df['sp_t'].apply(word_embed)
df['sp_embd']

## Search - Old

df.head()

def query_to_vec(query, column_used, embedder):
    token_query = query.split()
    clear_query = [token for token in token_query if token not in ['-0-', '-1-', '-2-', '-3-', '<UNK>']]
    embedded_tokens = np.array([embedder.wv.__getitem__(token) for token in clear_query if token in embedder.wv.key_to_index])
    if len(embedded_tokens > 0):
        return np.mean(embedded_tokens, axis=0)
    return [0 for i in range(len(df[column_used].dropna()[0]))]

def search(query, column_used, embedder):
    embedding_arrays = np.array(df[column_used].dropna().tolist())
    
    embed_query = query_to_vec(query, column_used, embedder)
    
    similarities = cosine_similarity(embedding_arrays, np.array(embed_query).reshape(1, -1))
    
    max_index = np.argmax(similarities)
    
    return df.at[max_index, 'poem_text']

query = 'جمال القمر'

print("Result using Morphological:")
print(search(query, 'morph_embd', morph_embedder))
print()

print("Result using Word:")
print(search(query, 'word_embd', word_embedder))
print()

print("Result using SentencePiece:")
print(search(query, 'sp_embd', sp_embedder))
print()

query = 'سيف'

print("Result using Morphological:")
print(search(query, 'morph_embd', morph_embedder))
print()

print("Result using Word:")
print(search(query, 'word_embd', word_embedder))
print()

print("Result using SentencePiece:")
print(search(query, 'sp_embd', sp_embedder))
print()

query = 'سيوف الوغى'

print("Result using Morphological:")
print(search(query, 'morph_embd', morph_embedder))
print()

print("Result using Word:")
print(search(query, 'word_embd', word_embedder))
print()

print("Result using SentencePiece:")
print(search(query, 'sp_embd', sp_embedder))
print()

## Bert + Word2Vec Embedding

df_by_line = pd.read_csv('all_poems_bert_processed.csv')
df_by_line.head()

def bert_embed(tokens_poem):
    clear_poem = [token for token in tokens_poem if not (stp.is_stop(token) or ('+' in token))]
    embedded_tokens = np.array([bert_embedder.wv.__getitem__(token) for token in clear_poem if token in bert_embedder.wv.key_to_index])
    return np.mean(embedded_tokens, axis=0)

df_by_line['bert_embd'] = df_by_line['bert_t'].apply(bert_embed)
df_by_line['bert_embd']

## Bert + Doc2Vec

def doc_to_vecs(string_vec):
    result_list = string_vec.strip("[]").split()
    return [float(item) for item in result_list]

df_by_line['doc_vecs'] = df_by_line['doc_vecs'].apply(doc_to_vecs)
df_by_line['doc_vecs'][0]

## Search - Doc2Vec

def query_to_docvec(query):
    token_query = query.split()
    return bert_doc_embedder.infer_vector(token_query)

def calc_similarity(doc_vector, query_vector):
    return cosine_similarity(np.array(doc_vector).reshape(1, -1), np.array(query_vector).reshape(1, -1))[0, 0]

def search_doc2vec(query, N=3):
    query_vec = query_to_docvec(query)
    similarities_series = df_by_line['doc_vecs'].apply(lambda doc_vec: calc_similarity(doc_vec, query_vec))
    max_indices = np.argpartition(similarities_series.values.tolist(), -N)[-N:]
    
    return df_by_line.loc[max_indices, 'poem_text']

query = 'جمال القمر'
print(search_doc2vec(query))

query = 'ثلوج الجبال'
print(search_doc2vec(query))

query = 'قتال الوغى'
print(search_doc2vec(query))

query = 'صوم رمضان'
print(search_doc2vec(query))

query = 'طول السفر'
print(search_doc2vec(query))

query = 'اشتياق الأهل'

print("Result using Bert:")
print(search_doc2vec(query))
print()

query = 'هجو السلطان'

print("Result using Bert:")
print(search_doc2vec(query))
print()

query = 'جور الخليفة'

print("Result using Bert:")
print(search_doc2vec(query))
print()

def query_to_docvec(query):
    token_query = query.split()
    return bert_doc_embedder.infer_vector(token_query)

def calc_similarity(doc_vector, query_vector):
    return cosine_similarity(doc_vector, query_vector)[0, 0]

def search_doc2vec_3(query, N=3):
    query_vec = np.array(query_to_docvec(query)).reshape(1, -1)
    similarities_series = df_by_line['doc_vecs'].apply(lambda doc_vec: calc_similarity(np.array(doc_vec).reshape(1, -1), query_vec))
    similarities_values = similarities_series.values.tolist()
    
    tri_way_similarity = [(similarities_values[i] + similarities_values[i-1] + similarities_values[i-2])/3 for i in range(len(similarities_values))]    
    max_indices = np.argpartition(tri_way_similarity, -N)[-N:]
    
    return df_by_line.loc[max_indices, 'poem_text']

query = 'صوم رمضان'

print("Result using Bert:")
print(search_doc2vec_3(query))
print()

query = 'ثلوج الجبال'

print("Result using Bert:")
print(search_doc2vec_3(query))
print()

query = 'شجاعة في أرض القتال'

print("Result using Bert:")
print(search_doc2vec_3(query))
print()

## Search - Word2Vec

def query_to_vec_bert(query):
    token_query = query.split()
    embedded_tokens = np.array([bert_embedder.wv.__getitem__(token) for token in token_query if token in bert_embedder.wv.key_to_index])
    if len(embedded_tokens > 0):
        return np.mean(embedded_tokens, axis=0)
    return [0 for i in range(len(['bert_embd'].dropna()[0]))]

def search_bert(query, N=3):
    embedding_arrays = np.array(df_by_line['bert_embd'].dropna().tolist())
    
    embed_query = query_to_vec_bert(query)
    
    similarities = cosine_similarity(embedding_arrays, np.array(embed_query).reshape(1, -1))
    
    max_indices = np.argpartition(similarities.flatten(), -N)[-N:]
    
    return df_by_line.loc[max_indices, 'poem_text']

query = 'صوم رمضان'
print(search_bert(query))

query = 'ثلوج الجبال'
print(search_bert(query))

query = 'شجاعة في أرض القتال'
print(search_bert(query))

query = 'طول السفر'
print(search_bert(query))
