==> 04_arabert_tokenization.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from itertools import islice
import arabicstopwords.arabicstopwords as stp

import pickle as pkl

import sentencepiece as sp

import tkseem as tk
import gensim

from arabert.preprocess import ArabertPreprocessor

model_name="bert-base-arabert"
arabert_prep = ArabertPreprocessor(model_name=model_name)

text = "ولن نبالغ إذا قلنا إن هاتف أو كمبيوتر المكتب في زمننا هذا ضروري"
arabert_prep.preprocess(text)

text = "هل غادر الشعراء من متردم أم هل عرفت الدار بعد توهم"
arabert_prep.preprocess(text)

text = "وتتابعت ذكرى الحبيب بخاطري والذكريات إذا جثمن جسام"
arabert_prep.preprocess(text)

text = "أثني علي بما علمت فإنني سمح مخالطتي إذا لم أظلم"
arabert_prep.preprocess(text)

df = pd.read_csv('data/apcd_clean.csv')
df['poem_text'] = df['البيت'].astype(str)
df.drop('البيت', inplace=True, axis=1)
df.head()

def tokenize_text_bert(text):
    tokenized_text = arabert_prep.preprocess(text).split()
    return [token for token in tokenized_text if '+' not in token]

df['bert_t'] = df.poem_text.apply(tokenize_text_bert)
df.head()

bert_model = gensim.models.Word2Vec(
    df['bert_t'],
    window=8,
    min_count=3,
    epochs=10,
    workers=8)

bert_model.wv.most_similar(positive='حرب')

bert_model.wv.most_similar(positive='قمر')

bert_model.wv.most_similar(positive='سلاح')

bert_model.wv.most_similar(positive='عيد')

bert_model.wv.most_similar(positive='رمضان')

bert_model.wv.most_similar(positive='صيام')

bert_model.wv.most_similar(positive='نهر')

bert_model.wv.most_similar(positive='ثكلى')

bert_model.wv.most_similar(positive='حب')

bert_model.wv.most_similar(positive='بكاء')

bert_model.wv.most_similar(positive='شرق')

bert_model.wv.most_similar(positive='تهجد')

bert_model.wv.most_similar(positive='عجوز')

bert_model.wv.most_similar(positive='شيطان')

bert_model.wv.most_similar(positive='سفيه')

bert_model.wv.most_similar(positive='ليت')

bert_model.wv.most_similar(positive='رائد')

bert_model.wv.most_similar(positive='شجاع')

bert_model.wv.most_similar(positive='عنتر')

bert_model.save('models/bert_word2vec_model.model')
df.to_csv('all_poems_bert_processed.csv', index=False)

df.head()

## Doc2Vec

from gensim.models.doc2vec import Doc2Vec, TaggedDocument
taggedData = [TaggedDocument(words=doc, tags=[str(i)]) for doc, i in zip(df['bert_t'], df.index)]

model = Doc2Vec(vector_size=150, min_count=4, epochs=20, workers=120, window=7, dm=1)
model

model.build_vocab(taggedData)

model.train(taggedData,
            total_examples=model.corpus_count,
            epochs=model.epochs)

#df['doc_vecs'] = df['bert_t'].apply(model.infer_vector)
df.head()

#df.to_csv('all_poems_bert_processed.csv', index=False)

model.save('models/bert_doc2vec_model_bow11.model')

prompt = 'وددت رؤية الياسمين'

clean_prompt = [token for token in arabert_prep.preprocess(prompt).split() if '+' not in token]
clean_prompt

inferred_vector = model.infer_vector(clean_prompt)
sims = model.dv.most_similar([inferred_vector], topn=500)
sims

array = np.array(sims)
df_ggl = pd.DataFrame(array, columns=['id', 'similarity'])
plt.plot(df_ggl.index, df_ggl['similarity'].astype(float))

df.loc[int(sims[2][0])]
