==> 06_bert_based_search.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from itertools import islice

import ast

import sentencepiece as sp

from sklearn.metrics.pairwise import cosine_similarity
import arabicstopwords.arabicstopwords as stp

from arabert.preprocess import ArabertPreprocessor

import tkseem as tk
import gensim

bert_embedder = gensim.models.Word2Vec.load("models/bert_word2vec_model.model")

model_name="bert-base-arabert"
arabert_prep = ArabertPreprocessor(model_name=model_name)

df = pd.read_csv('all_poems_bert_processed.csv').sample(100_000, random_state=42).reset_index()
df.head()

df['bert_t'] = df['bert_t'].apply(ast.literal_eval)
df['bert_t']

embedding_size = bert_embedder.wv.__getitem__('دجاج').shape[0]

def bert_embed(tokens_poem):
    weights = [0.3 if stp.is_stop(token) else 1 for token in tokens_poem if token in bert_embedder.wv.key_to_index]
    embedded_tokens = np.array([bert_embedder.wv.__getitem__(token) for token in tokens_poem if token in bert_embedder.wv.key_to_index])
    
    if(len(embedded_tokens) < 1):
        return np.zeros(embedding_size)
    
    if sum(weights) > 0:
        return np.average(embedded_tokens, axis=0, weights=weights)
    return np.mean(embedded_tokens, axis=0)

# TODO: Try hnaya tdire weighted embedding 3la 7sab l idf
# Note: Kayane TfidVectorizer f sklearn

# TODO: Sinon nsayo idea wa7do5ra:
#         Basically, mandirouch kml document embedding
#         Instead, njibo f kol document lmost similar words m3a l words nta3 l query nta3na
#         Wndiro weighted average 3la lsimilarities

df['bert_embd'] = df['bert_t'].apply(bert_embed)
df['bert_embd']

import faiss

arrays = df['bert_embd'].values.tolist()
embeddings = np.stack(arrays, axis=0)

dimension = embeddings[0].shape[0]
index = faiss.IndexFlatL2(embedding_size)
index.add(embeddings)
print(f'Total vectors in the index: {index.ntotal}')

query = 'بق بق نحن نغرق'
query_vector = bert_embed(query).reshape(1, -1)
distances, indices = index.search(query_vector, 20)
df.loc[indices.flatten()]

query = 'يا الهي'
query_vector = bert_embed(query).reshape(1, -1)
distances, indices = index.search(query_vector, 20)
df.loc[indices.flatten()]

### OLD CODE

query = 'خشيت موج البحر'

indices = search_bert(query, N=4)
for i, index in enumerate(indices):
    print((i+1), ': ', df.at[index, 'poem_text'])

query = 'الحرية ثمنها غالي'

verses = search_bert(query).values.tolist()
for i, verse in enumerate(verses):
    print((i+1), ': ', verse)

query = 'الحرية ثمنها غالي'

indices = search_bert(query, N=7)
for i, index in enumerate(indices):
    print((i+1), ': ', df.at[index, 'poem_text'])

query = 'عنترة'

verses = search_bert(query).values.tolist()
for i, verse in enumerate(verses):
    print((i+1), ': ', verse)

query = 'عنترة'

verses = search_avg_sim(query).values.tolist()
for i, verse in enumerate(verses):
    print((i+1), ': ', verse)

query = 'ليت شبابي يرجع'

indices = search_bert(query, N=4)
for i, index in enumerate(indices):
    print((i+1), ': ', df.at[index, 'poem_text'])

query = 'ليت شبابي يرجع'

verses = search_avg_sim(query).values.tolist()
for i, verse in enumerate(verses):
    print((i+1), ': ', verse)
