==> 07_conditional_generation.ipynb <==
# preparing the data

from torch.utils.data import DataLoader
import pandas as pd
from transformers import T5ForConditionalGeneration, T5Tokenizer
import torch

# 1. Preprocess Data

df = pd.read_csv('Arabic Poem Comprehensive Dataset (APCD).csv')
df.drop(columns=['البيت' , 'العصر' , 'الديوان' , 'القافية' , 'الشاعر' , 'البحر'], inplace=True)
df.dropna(subset=['الشطر الايمن', 'الشطر الايسر'], inplace=True)
poems = df.rename(columns={"الشطر الايمن": "prompt", "الشطر الايسر": "completion"})[:50000].to_dict(orient="records")

# 2. Create Dataset Class

class PoemDataset(torch.utils.data.Dataset):
    def __init__(self, poems, tokenizer, max_length):
        self.poems = poems
        self.tokenizer = tokenizer
        self.max_length = max_length

    def __len__(self):
        return len(self.poems)

    def __getitem__(self, idx):
        poem = self.poems[idx]
        prompt = poem["prompt"]
        completion = poem["completion"]

        # Tokenize the prompt and completion
        inputs = self.tokenizer.encode(prompt, add_special_tokens=True, max_length=self.max_length, truncation=True, padding='max_length', return_tensors="pt")
        labels = self.tokenizer.encode(completion, add_special_tokens=True, max_length=self.max_length, truncation=True, padding='max_length', return_tensors="pt")

        return {
            "input_ids": inputs.squeeze(0),  # Remove the batch dimension
            "labels": labels.squeeze(0),      # Remove the batch dimension
        }


#class PoemDataset(torch.utils.data.Dataset):
   # def __init__(self, poems):
       # self.poems = poems

  #  def __len__(self):
     #   return len(self.poems)

   # def __getitem__(self, idx):
       #poem = self.poems[idx]
        #prompt = poem["prompt"]
        #completion = poem["completion"]
        #encoded_data = tokenizer(prompt, completion, padding="max_length", truncation=True, return_tensors="pt")
        #return encoded_data



# 3. Model and Tokenizer

model_name = "bakrianoo/t5-arabic-small" 
model = T5ForConditionalGeneration.from_pretrained(model_name)
tokenizer = T5Tokenizer.from_pretrained(model_name)

# 4. Create Data Loader

max_length = 64 

# Create dataset and data loader
dataset = PoemDataset(poems, tokenizer, max_length)
data_loader = DataLoader(dataset, batch_size=32, shuffle=True)

data_loader

# 5. Data Preparation Function 
#def prepare_data(data):
    #encoded_data = tokenizer(data, padding="max_length", truncation=True, return_tensors="pt")
    #return encoded_data



# 6. Optimizer and Training Loop

num_epochs = 1
optimizer = torch.optim.Adam(model.parameters())

for epoch in range(num_epochs):
    for data in data_loader:
        inputs = data["input_ids"]
        labels = data["labels"]

        optimizer.zero_grad()
        outputs = model(input_ids=inputs, labels=labels)
        loss = outputs.loss
        loss.backward()
        optimizer.step()

# 7.  Generation Function 
def generate_poem(prompt, max_length=50, temperature=1.0, num_beams=5):
    input_ids = tokenizer.encode(prompt, return_tensors="pt")
    output = model.generate(
        input_ids=input_ids,
        max_length=max_length,
        temperature=temperature,
        num_beams=num_beams,
    )
    return tokenizer.decode(output[0], skip_special_tokens=True)


prompt = "وردة حمراء تتفتح"  
generated_poem = generate_poem(prompt)
print(generated_poem)

model.save('model_soumia.model')

torch.save(model.state_dict(), 'model_soumia_torch.model')
