==> 08_fine_tuning.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib
import seaborn as sns
from datasets import Dataset
import torch

from transformers import AutoTokenizer, AutoModelForCausalLM
df = pd.read_csv('Arabic Poem Comprehensive Dataset (APCD).csv', encoding_errors="ignore")
# Tokenize the text
tokenizer = AutoTokenizer.from_pretrained("CAMeL-Lab/bert-base-arabic-camelbert-mix")
model = AutoModelForCausalLM.from_pretrained("CAMeL-Lab/bert-base-arabic-camelbert-mix")
model_fine_tuned = AutoModelForCausalLM.from_pretrained("fine-tuned-bert-base-arabic-camelbert-mix", local_files_only=True)

model_fine_tuned

model_fine_tuned.push_to_hub('selma_model_20k_vv')

for col in df.columns:
  if col!="البيت":
    df.drop(col,axis=1,inplace=True)

df.dropna()
df['البيت']=df['البيت'].astype(str).values
df

from sklearn.model_selection import train_test_split

df_train, df_test = train_test_split(df.sample(100, random_state=42), test_size=0.2, random_state=42)
df_train=Dataset.from_pandas(df_train)
df_test=Dataset.from_pandas(df_test)

print("Train set shape:", df_train.shape)
print("Test set shape:", df_test.shape)

def tokenize_function(examples):
    return tokenizer(examples["البيت"], padding="max_length", truncation=True, max_length=200)

tokenized_datasets_train = df_train.map(tokenize_function, batched=True)
tokenized_datasets_test = df_test.map(tokenize_function, batched=True)

tokenized_datasets_train

### Fine-tuning CAMeL-Lab/bert-base-arabic-camelbert-mix Model

from transformers import TrainingArguments, Trainer

training_args = TrainingArguments(output_dir="selma_model_20k", push_to_hub=True, num_train_epochs=1, evaluation_strategy="epoch")

from transformers import DataCollatorForLanguageModeling
data_collator = DataCollatorForLanguageModeling(
    tokenizer=tokenizer, mlm=False#, mlm_probability=0.15
)

trainer = Trainer(
    model=model_fine_tuned,
    args=training_args,
    train_dataset=tokenized_datasets_train,
    eval_dataset=tokenized_datasets_test,
    data_collator=data_collator,
)

#torch.set_num_threads(10)

trainer.train()

trainer.save_model('model_selma_torch_20k_gg.model')

trainer.push_to_hub("selma_model_20k_gg")

tokenizer.push_to_hub("selma_tokenizer_20k")
