==> 09_fine_tuning_v2.ipynb <==
from transformers import AutoTokenizer, AutoModelForCausalLM

tokenizer = AutoTokenizer.from_pretrained("elgeish/gpt2-medium-arabic-poetry",is_split_into_words=True)
tokenizer.add_special_tokens({'pad_token': '[PAD]'})
model = AutoModelForCausalLM.from_pretrained("elgeish/gpt2-medium-arabic-poetry")
model.resize_token_embeddings(len(tokenizer))

# Downloading the data

import pandas as pd
from datasets import Dataset
df=pd.read_csv("./Arabic Poem Comprehensive Dataset (APCD).csv")
df['البيت']=df['البيت'].astype(str).values
df.dropna()

for col in df.columns:
  if col!="البيت":
    df.drop(col,axis=1,inplace=True)

df.shape

df_train=df[:50000]
df_test=df[50001:60000]
df_train=Dataset.from_pandas(df_train[:2000])

df_test=Dataset.from_pandas(df_test[:500])

df_train

def tokenize_function(examples):
    return tokenizer(examples["البيت"],padding='max_length', truncation=True, max_length=160)

tokenized_train = df_train.map(tokenize_function, batched=True)
tokenized_test=df_test.map(tokenize_function, batched=True)

print(tokenized_train[1]["البيت"])
print(tokenized_train[1]["input_ids"])
print(tokenized_train[1]["attention_mask"])


#@title
from IPython.display import HTML
from transformers import TrainingArguments, Trainer
#import huggingface_hub
#huggingface_hub.login('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')

HTML('')

training_args = TrainingArguments( evaluation_strategy="epoch",output_dir="raid-fine-tuned-gpt2-medium-arabic-poetry",push_to_hub_token="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", push_to_hub=True)

from transformers import DataCollatorForLanguageModeling
data_collator = DataCollatorForLanguageModeling(
    tokenizer=tokenizer, mlm=False#, mlm_probability=0.15
)

trainer = Trainer(
    model=model,
    args=training_args,
    train_dataset=tokenized_train,
    eval_dataset=tokenized_test,
    data_collator=data_collator
)

import torch
torch.set_num_threads(10)

trainer.train()

trainer.push_to_hub()

trainer.save_model('model_raid_3k')
