==> 10_sentence_transformer.ipynb <==
## Exploration

from sentence_transformers import SentenceTransformer
sentences = ["الجزائر في حالة جيدة", "أشعر بنعاس شديد"]

model = SentenceTransformer('medmediani/Arabic-KW-Mdel', device='cuda')
embeddings = model.encode(sentences)
print(embeddings)

import pandas as pd

df = pd.read_csv('data/data.csv')
df.rename({'bait_clean': 'البيت'}, axis=1, inplace=True)
df.head()

embeddings = model.encode(df['البيت'])
embeddings.shape

## Vector DB

import numpy as np
import faiss

# Number of vectors
num_vectors = embeddings.shape[0]
# Dimension of each vector
dimension = embeddings.shape[1]

# Initialize the index
index = faiss.IndexFlatL2(dimension)

# Add vectors to the index
index.add(embeddings)

# Print the number of vectors in the index
print(f'Total vectors in the index: {index.ntotal}')

faiss.write_index(index, 'vector_index.faiss')

## Old Code

verses = df.sample(50)['البيت'].values
verses

from scipy import spatial
import numpy as np

query = ['انقضاء الهوى']
query_embd = model.encode(query)[0]

def calc_similarity(arr):
    return spatial.distance.cosine(query_embd, arr)

embeddings = np.array(model.encode(verses))
similarities = np.apply_along_axis(calc_similarity, 1, embeddings)
similarities

verses[similarities.argsort()]

query = ['وجه كالبدر']
query_embd = model.encode(query)[0]
norm(query_embd)

## Search

df_sample = df.sample(5_000, random_state=42)
df_sample

def similarity(verse, query):
    embeddings = model.encode([verse, query])
    return 1 - spatial.distance.cosine(embeddings[0], embeddings[1])

query = 'غيث عظيم'
df_sample['similarity'] = df_sample['البيت'].apply(lambda verse: similarity(verse, query))
df_sample.sort_values(by=['similarity'], ascending=False)

query = 'ماء'
df_sample['similarity'] = df_sample['البيت'].apply(lambda verse: similarity(verse, query))
df_sample.sort_values(by=['similarity'], ascending=False)

query = 'عيون باكية'
df_sample['similarity'] = df_sample['البيت'].apply(lambda verse: similarity(verse, query))
df_sample.sort_values(by=['similarity'], ascending=False)

query = 'عقول ساهية'
df_sample['similarity'] = df_sample['البيت'].apply(lambda verse: similarity(verse, query))
df_sample.sort_values(by=['similarity'], ascending=False)

query = 'جيوش فانية'
df_sample['similarity'] = df_sample['البيت'].apply(lambda verse: similarity(verse, query))
df_sample.sort_values(by=['similarity'], ascending=False)

query = 'كنز ثمين'
df_sample['similarity'] = df_sample['البيت'].apply(lambda verse: similarity(verse, query))
df_sample.sort_values(by=['similarity'], ascending=False)
