==> 11_gpt_search.ipynb <==
## Import Model

from transformers import AutoModel, AutoTokenizer, set_seed
import torch

# Set random seed for reproducibility
set_seed(42)

# Model and tokenizer
model_name = "elgeish/gpt2-medium-arabic-poetry"
model = AutoModel.from_pretrained(model_name).to("cuda")
tokenizer = AutoTokenizer.from_pretrained(model_name)

# Input prompt
prompt = "مثلجات سلطع برجر"
input_ids = tokenizer.encode(prompt, return_tensors="pt").to("cuda")

# Ensure the model is in evaluation mode
model.eval()

# Get the model outputs
with torch.no_grad():
    outputs = model(input_ids)

# Extract hidden states from the last layer
hidden_states = outputs.last_hidden_state

# Calculate the sentence embedding as the mean of the token embeddings
sentence_embedding = torch.mean(hidden_states, dim=1).squeeze()

# Convert to numpy array (optional)
sentence_embedding = sentence_embedding.cpu().numpy()

print("Sentence embedding:", sentence_embedding)

import pandas as pd

df = pd.read_csv('data/apcd_clean.csv')
df.head()

from torch.utils.data import DataLoader, Dataset

set_seed(42)

class PromptDataset(Dataset):
    def __init__(self, prompts):
        self.prompts = prompts

    def __len__(self):
        return len(self.prompts)

    def __getitem__(self, idx):
        prompt = self.prompts[idx]
        input_ids = tokenizer.encode(prompt, return_tensors="pt").squeeze(0)
        return input_ids

# Create Dataset and DataLoader
dataset = PromptDataset(df.sample(10_000, random_state=42)['البيت'].values)
dataloader = DataLoader(dataset, batch_size=64, collate_fn=lambda x: torch.nn.utils.rnn.pad_sequence(x, batch_first=True))

# Ensure the model is in evaluation mode
model.eval()

# Function to extract embeddings
def extract_embeddings(dataloader, model):
    all_embeddings = []
    with torch.no_grad():
        for batch in dataloader:
            batch = batch.to("cuda")
            outputs = model(batch)
            hidden_states = outputs.last_hidden_state
            # Calculate the sentence embedding as the mean of the token embeddings
            batch_embeddings = torch.mean(hidden_states, dim=1)
            all_embeddings.append(batch_embeddings)
            
            del batch
            del outputs
            del hidden_states
            torch.cuda.empty_cache()
    # Concatenate all embeddings
    all_embeddings = torch.cat(all_embeddings, dim=0)
    return all_embeddings

# Extract embeddings
embeddings = extract_embeddings(dataloader, model)

# Convert to numpy array (optional)
embeddings = embeddings.cpu().numpy()

print("Embeddings shape:", embeddings.shape)
print(embeddings)

import faiss

dimension = embeddings[0].shape[0]
index = faiss.IndexFlatL2(dimension)
index.add(embeddings)
print(f'Total vectors in the index: {index.ntotal}')

original_indices = df.sample(10_000, random_state=42)['البيت'].index

prompts = [
    'الحماسة والفخر',
    'روعة البحر',
    'الأمل والتفاؤل',
    'معاني الحياة والموت',
    'وجدانية الحب العذري',
    'التأمل في مصائر الإنسان',
    'حياة العرب قبل الإسلام',
    'العدمية والبحث عن الذات',
    'الفضيلة والأخلاق',
    'السلام والتسامح',
    'العدالة والمساواة',
    'كيف تخاف الفقر والله رازقا',
    'التشرد والهجرة والبحث عن الوطن',
    'عيناك نازلتاالقلوب',
    ' كتب الدمع بخدي عهده',
    'لا يحمل الحقد من تعلو به الرتب',
    'أتعرف رسم الدار',
    'الأصدقاء مثل النجوم',
    ' أحلك الأوقات',
    'شعر بدوي',
    ' الثقة بالله',
    ' ذكريات الطفولة',
    'المرء يأمل أن يعيش',
    'الوطن والانتماء',
    ' الحرية والنضال',
    'قم للمعلم وفه التبجيلا',
    'وعذرتهم وعرفت ذنبي',
    'إليك يا أمي',
    'فأصبح بعد الحرب',
    'إذا حكم القضاء',
    'ليت الأحزان',
    'فرحي مثل طائر',
    'نار الغضب',
    'صوت المطر ينعش روحي',
    'تنير ليالي',
    'إِن الثعالب بالضحى',
    'صبور على مر الحوادث',
    'البحث عن الحقيقة رحلة لا نهاية لها',
    'مرآة الروح',
    'كلما نسينا دروس الماضي',
    'دع الأيام تفعل ما تشاء',
    'وقفت على الديار',
    'رمال الصحراء',
    'شموخ',
    'تفاؤل',
    'نصر',
    'حمامة',
    'ساذج'
]

f = open("output_gpt2.txt", 'w')

for prompt in prompts:
    f.write(prompt + "\n")

    set_seed(42)

    input_ids = tokenizer.encode(prompt, return_tensors="pt").to("cuda")

    with torch.no_grad():
        outputs = model(input_ids)

    # Extract hidden states from the last layer
    hidden_states = outputs.last_hidden_state

    # Calculate the sentence embedding as the mean of the token embeddings
    sentence_embedding = torch.mean(hidden_states, dim=1).squeeze()

    # Convert to numpy array (optional)
    query_vector = sentence_embedding.cpu().numpy().reshape(1, -1)
    
    distances, indices = index.search(query_vector, 50)
    df_prompt_result = df.loc[original_indices[indices.flatten()]]
    df_prompt_result['n'] = df_prompt_result['البيت'].apply(len)
    df_prompt_result = df_prompt_result.loc[df_prompt_result['n'] > 10]
    df_prompt_result.reset_index(inplace=True)
    
    f.write(df_prompt_result.at[0, 'البيت'] + "\n")
    f.write(df_prompt_result.at[1, 'البيت'] + "\n")
    f.write(df_prompt_result.at[2, 'البيت'] + "\n")
    
    f.write("\n")

f.close()

#faiss.write_index(index, 'vector_index.faiss')
