==> 13_last_word_predict.ipynb <==
import gensim

word_embedder = gensim.models.Word2Vec.load("models/bert_word2vec_model.model")
word_embedder

def replace_last_word(verse, kafiya):
    most_similar = word_embedder.wv.most_similar(positive=original.split()[-1], topn=100)
    possibilities = list(list(zip(*most_similar))[0])
    return next((s for s in possibilities if s.endswith(kafiya)), None)

original = 'وسريت في دجنة الليل ساهيا ** وجفوني كأنها شلال يسيل'
kafiya = 'ر'

replace_last_word(original, kafiya)
